+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v0.1.1
 - Added ability to start a new game
+- Keys D and Delete move a card to a top pile automatically if possible
 
 # v0.1.0
 - Initial release